% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifedata.R
\name{calculate}
\alias{calculate}
\title{Calculate Metrics}
\usage{
calculate(x, ...)
}
\arguments{
\item{x}{only defined for a fitted_life_data object}
}
\value{
returns numeric value
}
\description{
generic function for fitted_life_data only at this time
}
\details{
see calculate.fitted_life_data
}

