% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fit_data}
\alias{fit_data}
\title{Fit Distribution}
\usage{
fit_data(x, ...)
}
\arguments{
\item{x}{only defined for a lifedata object}
}
\description{
fits distribution to data
}
\details{
refer to \code{\link{fit_data.lifedata}}
}
\seealso{
\code{\link{fit_data.lifedata}}
}

