% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifedata.R
\name{hist.fitted_life_data}
\alias{hist.fitted_life_data}
\title{Histogram fitted_life_data Objects}
\usage{
\method{hist}{fitted_life_data}(x, type = "failure", theme = "base_r")
}
\value{
produces desired plot
}
\description{
failure/suspension histogram
}
\details{
see \code{\link{hist_lifedata}}
}
\seealso{
\code{\link{hist_lifedata}}
}

