% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifedata.R
\name{hist.lifedata}
\alias{hist.lifedata}
\title{Histogram lifedata Objects}
\usage{
\method{hist}{lifedata}(x, type = "failure", theme = "base_r")
}
\value{
produces desired plot
}
\description{
failure/suspension histogram
}
\details{
see \code{\link{hist_lifedata}}
}
\seealso{
\code{\link{hist_lifedata}}
}

