% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifedata.R
\name{hist_lifedata}
\alias{hist_lifedata}
\title{Histogram for Life Data Analysis}
\usage{
hist_lifedata(x, status, type, theme)
}
\arguments{
\item{x}{times from lifedata}

\item{status}{status from lifedata}

\item{type}{either failure or suspension}

\item{theme}{either base_r or weibull++}
}
\value{
produces desired plot
}
\description{
failure/suspension histogram
}
\details{
Produces histogram of either failure points or suspension points.
No graphical parameters can be passed
Theme inputs work similar to \code{\link{plot.fitted_life_data}}
}
\seealso{
\code{\link{hist}} \code{\link{plot.fitted_life_data}}
}

