% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifedata.R
\name{lifedata}
\alias{lifedata}
\title{Create lifedata Object.}
\usage{
lifedata(time, status, units = NULL)
}
\arguments{
\item{time}{numerical vector indicating time-to-event}

\item{status}{numerical vector indicating status. 0 for right censored. 1 for failure event.}

\item{units}{optional character to denote units of time}
}
\value{
a \code{lifedata} object
}
\description{
Create lifedata Object.
}
\examples{
lifedata(c(90.7, 114.8, 12.0, 144.35, 199.8), c(0,1,1,1,0), 'hours')
}

