% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifedata.R
\name{median_ranks.lifedata}
\alias{median_ranks.lifedata}
\title{Calculate Median Ranks}
\usage{
\method{median_ranks}{lifedata}(x, method = "inverse_F")
}
\arguments{
\item{x}{a lifedata object}

\item{method}{indicate calculation method. see details}
}
\value{
numeric vector corresponding to median ranks for failure data
}
\description{
calculates median ranks of lifedata object.
}
\details{
Median ranks are based on. 

By default, ranks are calculated based on solving 

For censored data, the mean order number of the failure points are determined first.
Median ranks are only calculated for failure points.
}

