% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifedata.R
\name{median_ranks}
\alias{median_ranks}
\title{Calculate Median Ranks}
\usage{
median_ranks(x, ...)
}
\arguments{
\item{x}{only method defined for lifedata objects}
}
\description{
s3 generic
}
\details{
refer to median_ranks.lifedata
}

