% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifedata.R
\name{pieplot.fitted_life_data}
\alias{pieplot.fitted_life_data}
\title{Pie Chart fitted_life_data Objects}
\usage{
\method{pieplot}{fitted_life_data}(x, theme = "base_r", ...)
}
\value{
produces desired plot
}
\description{
failure/suspension pie chart
}
\details{
see \code{\link{pieplot_lifedata}}
}
\seealso{
\code{\link{pieplot_lifedata}}
}

