% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifedata.R
\name{pieplot.lifedata}
\alias{pieplot.lifedata}
\title{Pie Chart lifedata Objects}
\usage{
\method{pieplot}{lifedata}(x, theme = "base_r", ...)
}
\value{
produces desired plot
}
\description{
failure/suspension pie chart
}
\details{
see \code{\link{pieplot_lifedata}}
}
\seealso{
\code{\link{pieplot_lifedata}}
}

