% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifedata.R
\name{pieplot_lifedata}
\alias{pieplot_lifedata}
\title{Pie Charts for Life Data Analysis}
\usage{
pieplot_lifedata(x, theme, ...)
}
\arguments{
\item{x}{status values passed from generics}

\item{theme}{either 'base_r' or 'weibull++'}
}
\value{
produces desired plot
}
\description{
failure/suspension pie chart
}
\details{
Graphical parameters can be passed when theme = 'base_r'.
}
\seealso{
\code{\link{pie}}
}

