% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifedata.R
\name{plot.fitted_life_data}
\alias{plot.fitted_life_data}
\title{Plotting fitted_life_data Objects}
\usage{
\method{plot}{fitted_life_data}(x, type = "probability", theme = "base_r")
}
\arguments{
\item{x}{a fitted_life_data object}

\item{type}{type of plot desired. default to 'probability'. see details}

\item{theme}{character indicating style. only 'base_r' and 'weibull++' are available. see details.}
}
\value{
produces desired plot
}
\description{
Provides various plots of fitted distribution
}
\details{
Values for type input
\tabular{ll}{
probability   \tab linearized scale plot. distribution specific.\cr
failure \tab probability of failure over time i.e. CDF\cr
reliability \tab reliability over time i.e. 1-CDF \cr
pdf \tab fitted probability density function \cr
failure rate \tab fitted failure rate i.e. hazard function \cr
}
First 3 plots will included points corresponding to the 
median ranks of the failure data.

By convention, in linearized scales, the exponential distribution plots the reliability while 
Weibull plots the unreliability.

When theme = 'weibull++', the plots styling is automatically set to mirror Weibull++.
By default, theme = 'base_r' and plots a very simple plot.
At this time, no graphical parameters can be passed.
}

