% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifedata.R
\name{timeline_lifedata}
\alias{timeline_lifedata}
\title{Timeline Charts for Life Data Analysis}
\usage{
timeline_lifedata(x, theme, ...)
}
\arguments{
\item{x}{lifedata object}

\item{theme}{either 'base_r' or 'weibull++'}
}
\value{
produces desired plot
}
\description{
Timeline chart showing times
}
\details{
Chart is not designed for large number of data points.
Graphical parameters can be passed when theme = 'base_r'.
}
\seealso{
\code{\link{pie}}
}

