% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifedata.R
\name{fit_data.lifedata}
\alias{fit_data.lifedata}
\title{Fit Distribution to lifedata Object}
\usage{
\method{fit_data}{lifedata}(x, dist = "weibull")
}
\arguments{
\item{x}{a lifedata object}

\item{dist}{character indicating distribution. only exponential and weibull are available}
}
\value{
A fitted_life_data object.
Object includes original input lifedata object, the distribution fitted, log likelihood, parameter estimates.
If all data points are complete, a chi-square goodness-of-fit output is included
}
\description{
Fits a distribution to lifedata object.
Only MLE are done and only weibull and exponential distributions are available
}
\details{
For exponential distributions, uses analytic solution.
For Weibull distributions, use optim function to find MLE.
}

