% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifedata.R
\name{plot.fitted_life_data}
\alias{plot.fitted_life_data}
\title{Plotting fitted_life_data Objects}
\usage{
\method{plot}{fitted_life_data}(x, type = "probability", theme = "base_r",
  line_par = NULL, point_par = NULL, ...)
}
\arguments{
\item{x}{a fitted_life_data object}

\item{type}{type of plot desired. default to 'probability'. see details}

\item{theme}{character indicating style. only 'base_r' and 'weibull++' are available. see details.}

\item{line_par}{a list containing graphical parameters for lines}

\item{point_par}{a list contianing graphical parameters for points}
}
\value{
produces desired plot
}
\description{
Provides various plots of fitted distribution
}
\details{
Values for type input
\tabular{ll}{
probability   \tab linearized scale plot. distribution specific.\cr
failure \tab probability of failure over time i.e. CDF\cr
reliability \tab reliability over time i.e. 1-CDF \cr
pdf \tab fitted probability density function \cr
failure rate \tab fitted failure rate i.e. hazard function \cr
}
First 3 plots will included points corresponding to the 
median ranks of the failure data.

By convention, in linearized scales, the exponential distribution plots the reliability while 
Weibull plots the unreliability.

When theme = 'weibull++', the plots styling is automatically set to mirror Weibull++.
By default, theme = 'base_r' and plots a very simple plot.

Graphical parameters can be passed when theme = 'base_r' but not for others.
When passing graphical parameters, those for lines should be passed as a list in line_par.
Those for points should be passed as a list in point_par.
All others e.g., xlab, should be passed generically (...)
}

